% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_regression.R
\name{Linear_regression}
\alias{Linear_regression}
\alias{linear_regression}
\title{Regressao Linear}
\usage{
linear_regression(data, resposta, preditor, conf.level = 0.95)
}
\arguments{
\item{data}{Dataframe contendo os dados de entrada.}

\item{resposta}{Nome da coluna que contem a variavel resposta (dependente).}

\item{preditor}{Lista de nomes das colunas que contem as variaveis preditoras (independentes).}

\item{conf.level}{Nivel de confianca para os intervalos de confianca, padrao e 0.95.}
}
\value{
Um objeto com os coeficientes, residuos, valores preditos, p-valores e intervalos de confianca.
}
\description{
Ajusta um modelo de regressao linear com base em um dataframe fornecido. A funcao retorna um
objeto contendo coeficientes, residuos, valores preditos, p-valores e intervalos de confianca.
}
\examples{
library (primeiropacote)
modelo <- linear_regression(simulated_data, "Y", c("X1", "X2"), conf.level = 0.95)
summary(modelo)  # Resumo do modelo ajustado
}
