% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minha_primeira_funcao.R
\name{ola}
\alias{ola}
\title{Retorna Mensagem.}
\usage{
ola()
}
\value{
Uma string contendo a mensagem "Hello, World!".
}
\description{
Uma funcao simples que retorna uma mensagem de saudacao.
}
\examples{
ola()  # Retorna "Hello, World!"
}
