% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{plot_regression}
\alias{plot_regression}
\title{Gráfico de Valores Preditos vs Observados}
\usage{
plot_regression(model)
}
\arguments{
\item{model}{Resultado da função linear_regression.}
}
\description{
Gráfico de Valores Preditos vs Observados
}
