% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Conjunto de Dados Simulados}
\format{
Um dataframe com 100 observações e 3 variáveis.
\describe{
  \item{X1}{Variável preditora 1.}
  \item{X2}{Variável preditora 2.}
  \item{Y}{Variável resposta.}
}
}
\usage{
simulated_data
}
\description{
Um dataframe contendo dados simulados para ilustração.
}
\keyword{datasets}
