% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Regressão Linear}
\usage{
linear_regression(data, resposta, preditor, conf.level = 0.95)
}
\arguments{
\item{data}{Dataframe.}

\item{resposta}{Nome da coluna de variável resposta.}

\item{preditor}{Lista de nomes das colunas de variáveis preditoras.}

\item{conf.level}{Nível de confiança para os intervalos de confiança.}
}
\value{
Um objeto com os coeficientes, resíduos, valores preditos, p-valores e intervalos de confiança.
}
\description{
Ajusta um modelo de regressão linear.
}
