% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{plot_regression_predict}
\alias{plot_regression_predict}
\title{Gráfico de Valores Reais vs Preditos}
\usage{
plot_regression_predict(model)
}
\arguments{
\item{model}{Resultado da função de regressão linear, contendo valores preditos e resíduos.}
}
\description{
Gráfico de Valores Reais vs Preditos
}
