% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{plot_regression_residuos}
\alias{plot_regression_residuos}
\title{Gráfico de Valores Preditos vs Resíduos}
\usage{
plot_regression_residuos(model)
}
\arguments{
\item{model}{Resultado da função linear_regression.}
}
\description{
Gráfico de Valores Preditos vs Resíduos
}
