% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{predict_new}
\alias{predict_new}
\title{Predição}
\usage{
predict_new(model, new_data)
}
\arguments{
\item{model}{Resultado da função linear_regression.}

\item{new_data}{Dataframe com novas observações.}
}
\value{
Valores preditos para as novas observações.
}
\description{
Predição
}
