% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Conjunto de Dados Simulados}
\format{
Um dataframe com 100 observacoes e 3 variaveis.
\describe{
  \item{X1}{Variavel preditora 1, gerada aleatoriamente a partir de uma distribuicao normal.}
  \item{X2}{Variavel preditora 2, gerada aleatoriamente a partir de uma distribuicao normal.}
  \item{Y}{Variavel resposta, gerada aleatoriamente a partir de uma distribuicao normal.}
}
}
\usage{
simulated_data
}
\description{
Um dataframe contendo dados simulados para ilustracao. Este conjunto e util para testar
funcoes de regressao linear e visualizacao.
}
\examples{
head(simulated_data)  # Visualizar as primeiras linhas do conjunto de dados simulado
summary(simulated_data)  # Resumo estatistico das variaveis
}
\keyword{datasets}
